sql: fix ephemeral table next rowid generation
Before this commit, next rowid was retrieved from an ephemeral table
using `index_max`. That led to wrong behavior because the index is not
sorted by rowid and `index_max` doesn`t return a tuple with the greatest
rowid.

New implementation stores next `rowid` value in BtCursor, and increments
it after each insert to the ephemeral space.

Extra refactoring:
 * `nKey` attribute is deleted from BtCursor, because it is not used
anymore.

Closes #3297