wal: rollback vclock on write failure
In order to determine whether we need to rebootstrap the instance on
startup, we need to know its vclock. To find it out, we scan the last
xlog file before proceeding to local recovery, but this means in case
rebootstrap is not required we scan the last xlog twice, which is
sub-optimal. To avoid double scan, we can create a new empty xlog before
shutting down the server and reopen it after restart. However, since we
promote WAL writer vclock even if xlog write fails, there will be an LSN
gap between the last xlog and the one created on shutdown in case we
failed to write last few records. To avoid that, let's rollback WAL
writer vclock if write fails. BTW this will make it consistent with
replicaset vclock - see commit 3c4bac715960a ("Follow vclock only for
success wal writes").